+++ src/nativeAd/NativeAdView.tsx
@@ -1,24 +1,14 @@
 import * as React from 'react';
 import { forwardRef, useContext, useImperativeHandle, useRef, useState, useEffect, useCallback } from 'react';
-import { NativeModules, requireNativeComponent, UIManager, findNodeHandle, View } from 'react-native';
-import type { ViewProps } from 'react-native';
+import { findNodeHandle, View } from 'react-native';
+import type { NativeSyntheticEvent, ViewProps } from 'react-native';
+import AppLovinMAX from '../specs/NativeAppLovinMAXModule';
+import NativeAdViewComponent, { Commands } from '../specs/AppLovinMAXNativeAdViewNativeComponent';
+import type { AdInfoEvent, AdLoadFailedEvent } from '../specs/AppLovinMAXNativeAdViewNativeComponent';
 import { NativeAdViewContext, NativeAdViewProvider } from './NativeAdViewProvider';
-import type { AdInfo, AdLoadFailedInfo } from '../types/AdInfo';
-import type { AdNativeEvent } from '../types/AdEvent';
-import type { NativeAd } from '../types/NativeAd';
+import type { NativeAdViewContextType } from './NativeAdViewProvider';
 import type { NativeAdViewHandler, NativeAdViewProps } from '../types/NativeAdViewProps';
-import type { NativeAdViewType, NativeAdViewContextType } from './NativeAdViewProvider';
-
-const { AppLovinMAX } = NativeModules;
-
-type NativeAdViewNativeEvents = {
-    onAdLoadedEvent(event: { nativeEvent: { nativeAd: NativeAd; adInfo: AdInfo } }): void;
-    onAdLoadFailedEvent(event: AdNativeEvent<AdLoadFailedInfo>): void;
-    onAdClickedEvent(event: AdNativeEvent<AdInfo>): void;
-    onAdRevenuePaidEvent(event: AdNativeEvent<AdInfo>): void;
-};
-
-const NativeAdViewComponent = requireNativeComponent<NativeAdViewProps & ViewProps & NativeAdViewNativeEvents>('AppLovinMAXNativeAdView');
+import { makeExtraParametersArray, makeLocalExtraParametersArray } from '../Utils';
 
 /**
  * The {@link NativeAdView} component that you use building a native ad. This loads a native ad and
@@ -61,15 +51,13 @@ export const NativeAdView = forwardRef<NativeAdViewHandler, NativeAdViewProps & 
     const [isInitialized, setIsInitialized] = useState<boolean>(false);
 
     useEffect(() => {
-        const checkInitialization = async () => {
+        (async () => {
             const result = await AppLovinMAX.isInitialized();
             setIsInitialized(result);
             if (!result) {
                 console.warn('NativeAdView is mounted before the initialization of the AppLovin MAX React Native module.');
             }
-        };
-
-        checkInitialization();
+        })();
     }, []);
 
     // Avoid rendering the NativeAdView if the SDK is not initialized
@@ -84,76 +72,87 @@ export const NativeAdView = forwardRef<NativeAdViewHandler, NativeAdViewProps & 
     );
 });
 
+const handleNativeAdViewEvent = <T extends AdInfoEvent | AdLoadFailedEvent>(event: NativeSyntheticEvent<T>, callback?: (adInfo: T) => void) => {
+    if (!callback) return;
+
+    let adInfo: any = { ...event.nativeEvent };
+
+    if ('nativeAd' in event.nativeEvent) {
+        adInfo.nativeAd = {
+            ...event.nativeEvent.nativeAd,
+            isIconImageAvailable: event.nativeEvent.nativeAd?.isIconImageAvailable ?? false,
+            isOptionsViewAvailable: event.nativeEvent.nativeAd?.isOptionsViewAvailable ?? false,
+            isMediaViewAvailable: event.nativeEvent.nativeAd?.isMediaViewAvailable ?? false,
+        };
+    }
+
+    callback(adInfo);
+};
+
 const NativeAdViewImpl = forwardRef<NativeAdViewHandler, NativeAdViewProps & ViewProps>(function NativeAdViewImpl(
     { adUnitId, placement, customData, extraParameters, localExtraParameters, onAdLoaded, onAdLoadFailed, onAdClicked, onAdRevenuePaid, children, style, ...otherProps },
     ref
 ) {
     // Context provides functions to manage native ad and native ad view state
-    const { setNativeAd, setNativeAdView } = useContext(NativeAdViewContext) as NativeAdViewContextType;
+    const { titleRef, advertiserRef, bodyRef, callToActionRef, imageRef, optionViewRef, mediaViewRef, setNativeAd } = useContext(NativeAdViewContext) as NativeAdViewContextType;
 
-    const nativeAdViewRef = useRef<NativeAdViewType | null>(null);
+    const nativeAdViewRef = useRef<React.ElementRef<typeof NativeAdViewComponent> | undefined>();
 
     // Load a new ad
     const loadAd = useCallback(() => {
-        const nativeAdView = nativeAdViewRef.current;
-        if (nativeAdView) {
-            const viewManagerConfig = UIManager.getViewManagerConfig('AppLovinMAXNativeAdView');
-            if (viewManagerConfig?.Commands && typeof viewManagerConfig.Commands.loadAd === 'number') {
-                UIManager.dispatchViewManagerCommand(findNodeHandle(nativeAdView), viewManagerConfig.Commands.loadAd, []);
-            }
-        }
+        nativeAdViewRef.current && Commands.loadAd(nativeAdViewRef.current);
     }, []);
 
     useImperativeHandle(ref, () => ({ loadAd }), [loadAd]);
 
-    // Save the DOM element reference
-    const saveElement = useCallback(
-        (element: NativeAdViewType | null) => {
-            if (element) {
-                nativeAdViewRef.current = element;
-                setNativeAdView(element);
-            }
-        },
-        [setNativeAdView]
-    );
+    const updateAssetView = useCallback((assetViewRef: React.RefObject<View>, type: string) => {
+        if (!nativeAdViewRef.current || !assetViewRef.current) return;
+
+        const node = findNodeHandle(assetViewRef.current);
+        if (node) {
+            Commands.updateAssetView(nativeAdViewRef.current, node, type);
+        }
+    }, []);
 
     const onAdLoadedEvent = useCallback(
-        (event: { nativeEvent: { nativeAd: NativeAd; adInfo: AdInfo } }) => {
-            setNativeAd(event.nativeEvent.nativeAd);
-            onAdLoaded?.(event.nativeEvent.adInfo);
+        (event: NativeSyntheticEvent<AdInfoEvent>) => {
+            const nativeAdImpl = event.nativeEvent.nativeAdImpl;
+
+            if (nativeAdImpl) {
+                setNativeAd({ ...nativeAdImpl });
+
+                if (nativeAdImpl.title) updateAssetView(titleRef, 'TitleView');
+                if (nativeAdImpl.advertiser) updateAssetView(advertiserRef, 'AdvertiserView');
+                if (nativeAdImpl.body) updateAssetView(bodyRef, 'BodyView');
+                if (nativeAdImpl.callToAction) updateAssetView(callToActionRef, 'CallToActionView');
+                if (nativeAdImpl.url || nativeAdImpl.image || nativeAdImpl.imageSource) updateAssetView(imageRef, 'IconView');
+                if (nativeAdImpl.isOptionsViewAvailable) updateAssetView(optionViewRef, 'OptionsView');
+                if (nativeAdImpl.isMediaViewAvailable) updateAssetView(mediaViewRef, 'MediaView');
+            }
+
+            Commands.renderNativeAd(nativeAdViewRef.current!);
+
+            handleNativeAdViewEvent(event, onAdLoaded);
         },
-        [onAdLoaded, setNativeAd]
+        [setNativeAd, onAdLoaded, updateAssetView, titleRef, advertiserRef, bodyRef, callToActionRef, imageRef, optionViewRef, mediaViewRef]
     );
 
-    const onAdLoadFailedEvent = useCallback(
-        (event: AdNativeEvent<AdLoadFailedInfo>) => {
-            onAdLoadFailed?.(event.nativeEvent);
-        },
-        [onAdLoadFailed]
-    );
+    const onAdLoadFailedEvent = useCallback((event: NativeSyntheticEvent<AdLoadFailedEvent>) => handleNativeAdViewEvent(event, onAdLoadFailed), [onAdLoadFailed]);
 
-    const onAdClickedEvent = useCallback(
-        (event: AdNativeEvent<AdInfo>) => {
-            onAdClicked?.(event.nativeEvent);
-        },
-        [onAdClicked]
-    );
+    const onAdClickedEvent = useCallback((event: NativeSyntheticEvent<AdInfoEvent>) => handleNativeAdViewEvent(event, onAdClicked), [onAdClicked]);
 
-    const onAdRevenuePaidEvent = useCallback(
-        (event: AdNativeEvent<AdInfo>) => {
-            onAdRevenuePaid?.(event.nativeEvent);
-        },
-        [onAdRevenuePaid]
-    );
+    const onAdRevenuePaidEvent = useCallback((event: NativeSyntheticEvent<AdInfoEvent>) => handleNativeAdViewEvent(event, onAdRevenuePaid), [onAdRevenuePaid]);
 
     return (
         <NativeAdViewComponent
-            ref={saveElement}
+            ref={(element) => (nativeAdViewRef.current = element ?? undefined)}
             adUnitId={adUnitId}
             placement={placement}
             customData={customData}
-            extraParameters={extraParameters}
-            localExtraParameters={localExtraParameters}
+            extraParameters={makeExtraParametersArray(extraParameters)}
+            strLocalExtraParameters={makeLocalExtraParametersArray(localExtraParameters, 'string')}
+            numLocalExtraParameters={makeLocalExtraParametersArray(localExtraParameters, 'number')}
+            boolLocalExtraParameters={makeLocalExtraParametersArray(localExtraParameters, 'bool')}
             onAdLoadedEvent={onAdLoadedEvent}
             onAdLoadFailedEvent={onAdLoadFailedEvent}
             onAdClickedEvent={onAdClickedEvent}