+++ src/nativeAd/NativeAdViewComponents.tsx
@@ -1,77 +1,50 @@
 import * as React from 'react';
-import { useContext, useRef, useEffect, useCallback, useMemo } from 'react';
-import { findNodeHandle, Text, Image, View, TouchableOpacity, StyleSheet, Platform } from 'react-native';
-import type { ViewProps, ImageProps, TextStyle, StyleProp, TextProps } from 'react-native';
+import { useContext, useMemo } from 'react';
+import { Text, Image, View, TouchableOpacity, StyleSheet, Platform } from 'react-native';
+import type { ViewProps, ImageProps, TextStyle, TextProps } from 'react-native';
 import { NativeAdViewContext } from './NativeAdViewProvider';
-import type { NativeAd } from '../types/NativeAd';
-
-// Custom hook to handle setting native ad view properties and return nativeAd
-const useNativeAdViewProps = (nativeAdProp: keyof NativeAd, ref: React.RefObject<any>, nativePropKey: string) => {
-    const { nativeAd, nativeAdView } = useContext(NativeAdViewContext);
-
-    const setNativeProps = useCallback(() => {
-        if (!nativeAd[nativeAdProp] || !ref.current) return;
-        nativeAdView?.setNativeProps({
-            [nativePropKey]: findNodeHandle(ref.current),
-        });
-    }, [nativeAd, nativeAdProp, nativeAdView, ref, nativePropKey]);
-
-    useEffect(() => {
-        setNativeProps();
-    }, [setNativeProps]);
-
-    return nativeAd;
-};
 
 export const TitleView = (props: TextProps) => {
-    const titleRef = useRef<Text | null>(null);
-    const nativeAd = useNativeAdViewProps('title', titleRef, 'titleView');
-
+    const { titleRef, nativeAd } = useContext(NativeAdViewContext);
     return (
         <Text {...props} ref={titleRef}>
-            {nativeAd.title || null}
+            {nativeAd.title ?? ''}
         </Text>
     );
 };
 
 export const AdvertiserView = (props: TextProps) => {
-    const advertiserRef = useRef<Text | null>(null);
-    const nativeAd = useNativeAdViewProps('advertiser', advertiserRef, 'advertiserView');
-
+    const { advertiserRef, nativeAd } = useContext(NativeAdViewContext);
     return (
         <Text {...props} ref={advertiserRef}>
-            {nativeAd.advertiser || null}
+            {nativeAd.advertiser ?? ''}
         </Text>
     );
 };
 
 export const BodyView = (props: TextProps) => {
-    const bodyRef = useRef<Text | null>(null);
-    const nativeAd = useNativeAdViewProps('body', bodyRef, 'bodyView');
-
+    const { bodyRef, nativeAd } = useContext(NativeAdViewContext);
     return (
         <Text {...props} ref={bodyRef}>
-            {nativeAd.body || null}
+            {nativeAd.body ?? ''}
         </Text>
     );
 };
 
 export const CallToActionView = (props: TextProps) => {
-    const callToActionRef = useRef<Text | null>(null);
-    const nativeAd = useNativeAdViewProps('callToAction', callToActionRef, 'callToActionView');
-
+    const { callToActionRef, nativeAd } = useContext(NativeAdViewContext);
     // TouchableOpacity disables clicking on certain Android devices.
     if (Platform.OS === 'android') {
         return (
             <Text {...props} ref={callToActionRef}>
-                {nativeAd.callToAction || null}
+                {nativeAd.callToAction ?? ''}
             </Text>
         );
     } else {
         return (
             <TouchableOpacity>
                 <Text {...props} ref={callToActionRef}>
-                    {nativeAd.callToAction || null}
+                    {nativeAd.callToAction ?? ''}
                 </Text>
             </TouchableOpacity>
         );
@@ -79,39 +52,29 @@ export const CallToActionView = (props: TextProps) => {
 };
 
 export const IconView = (props: Omit<ImageProps, 'source'>) => {
-    const imageRef = useRef<Image | null>(null);
-    const nativeAd = useNativeAdViewProps('image', imageRef, 'iconView');
-
-    return nativeAd.url ? (
-        <Image {...props} ref={imageRef} source={{ uri: nativeAd.url }} />
-    ) : nativeAd.image ? (
-        <Image {...props} ref={imageRef} source={0} />
-    ) : (
-        <View {...props} />
-    );
+    const { imageRef, nativeAd } = useContext(NativeAdViewContext);
+    const defaultIcon = require('./img/blank_icon.png');
+    const imageSource =
+        (nativeAd?.url && { uri: nativeAd.url }) || (nativeAd.image ? defaultIcon : nativeAd.imageSource ? { uri: `data:image/jpeg;base64,${nativeAd.imageSource}` } : defaultIcon);
+    return <Image {...props} ref={imageRef} source={imageSource} />;
 };
 
 export const OptionsView = (props: ViewProps) => {
-    const viewRef = useRef<View | null>(null);
-    useNativeAdViewProps('isOptionsViewAvailable', viewRef, 'optionsView');
-
-    return <View {...props} ref={viewRef} />;
+    const { optionViewRef } = useContext(NativeAdViewContext);
+    return <View {...props} ref={optionViewRef} />;
 };
 
 export const MediaView = (props: ViewProps) => {
-    const viewRef = useRef<View | null>(null);
-    useNativeAdViewProps('isMediaViewAvailable', viewRef, 'mediaView');
-
-    return <View {...props} ref={viewRef} />;
+    const { mediaViewRef } = useContext(NativeAdViewContext);
+    return <View {...props} ref={mediaViewRef} />;
 };
 
 export const StarRatingView = (props: ViewProps) => {
     const { style, ...restProps } = props;
-
     const maxStarCount = 5;
-    const starColor = StyleSheet.flatten((style as StyleProp<TextStyle>) || {}).color ?? '#ffe234';
-    const starSize = StyleSheet.flatten((style as StyleProp<TextStyle>) || {}).fontSize ?? 10;
-
+    const starTextStyle = StyleSheet.flatten(style) as TextStyle;
+    const starColor = starTextStyle.color ?? '#ffe234';
+    const starSize = starTextStyle.fontSize ?? 10;
     const { nativeAd } = useContext(NativeAdViewContext);
 
     // Memoize the star rendering process