+++ src/nativeAd/NativeAdViewProvider.tsx
@@ -1,17 +1,22 @@
 import * as React from 'react';
-import { useState, createContext } from 'react';
+import { useState, createContext, useRef } from 'react';
 import type { ReactNode } from 'react';
-import type { NativeMethods } from 'react-native';
+import type { Image, NativeMethods, Text, View } from 'react-native';
 import type { NativeAd } from '../types/NativeAd';
-import type { NativeAdViewProps } from '../types/NativeAdViewProps';
+import type { NativeProps } from '../specs/AppLovinMAXNativeAdViewNativeComponent';
 
-export type NativeAdViewType = React.Component<NativeAdViewProps> & NativeMethods;
+export type NativeAdViewType = React.Component<NativeProps> & NativeMethods;
 
 export type NativeAdViewContextType = {
+    titleRef: React.RefObject<React.ElementRef<typeof Text>>;
+    advertiserRef: React.RefObject<React.ElementRef<typeof Text>>;
+    bodyRef: React.RefObject<React.ElementRef<typeof Text>>;
+    callToActionRef: React.RefObject<React.ElementRef<typeof Text>>;
+    imageRef: React.RefObject<React.ElementRef<typeof Image>>;
+    optionViewRef: React.RefObject<React.ElementRef<typeof View>>;
+    mediaViewRef: React.RefObject<React.ElementRef<typeof View>>;
     nativeAd: NativeAd;
-    nativeAdView: NativeAdViewType | null;
     setNativeAd: React.Dispatch<React.SetStateAction<NativeAd>>;
-    setNativeAdView: React.Dispatch<React.SetStateAction<NativeAdViewType | null>>;
 };
 
 const defaultNativeAd: NativeAd = {
@@ -20,24 +25,40 @@ const defaultNativeAd: NativeAd = {
 };
 
 export const NativeAdViewContext = createContext<NativeAdViewContextType>({
+    titleRef: { current: null },
+    advertiserRef: { current: null },
+    bodyRef: { current: null },
+    callToActionRef: { current: null },
+    imageRef: { current: null },
+    optionViewRef: { current: null },
+    mediaViewRef: { current: null },
     nativeAd: defaultNativeAd,
-    nativeAdView: null,
     setNativeAd: () => {},
-    setNativeAdView: () => {},
 });
 
 export const NativeAdViewProvider: React.FC<{ children: ReactNode }> = ({ children }) => {
+    const titleRef = useRef<React.ElementRef<typeof Text> | null>(null);
+    const advertiserRef = useRef<React.ElementRef<typeof Text> | null>(null);
+    const bodyRef = useRef<React.ElementRef<typeof Text> | null>(null);
+    const callToActionRef = useRef<React.ElementRef<typeof Text> | null>(null);
+    const imageRef = useRef<React.ElementRef<typeof Image> | null>(null);
+    const optionViewRef = useRef<React.ElementRef<typeof View> | null>(null);
+    const mediaViewRef = useRef<React.ElementRef<typeof View> | null>(null);
     const [nativeAd, setNativeAd] = useState<NativeAd>(defaultNativeAd);
-    const [nativeAdView, setNativeAdView] = useState<NativeAdViewType | null>(null);
 
     const providerValue = React.useMemo(
         () => ({
+            titleRef,
+            advertiserRef,
+            bodyRef,
+            callToActionRef,
+            imageRef,
+            optionViewRef,
+            mediaViewRef,
             nativeAd,
-            nativeAdView,
             setNativeAd,
-            setNativeAdView,
         }),
-        [nativeAd, nativeAdView]
+        [nativeAd]
     );
 
     return <NativeAdViewContext.Provider value={providerValue}>{children}</NativeAdViewContext.Provider>;